Working save on Wipe with added event
It works while an event set in the pipeline is not set, and is only set when the program is shutting down.
It then waits for the pipeline to have a new image ready, and then it starts the comoutations.

If the window is rezised it saves, if there is some deletion it will wait for the whole thing to have been deleted, before saving the last snapshot.